feat: 添加营销模块路由和侧边栏接口
- 新增客户群发、客户继承、群群发、任务宝营销模块路由
- 添加侧边栏客户标签、跟进状态、媒体分组接口
- 修复企业列表corpId字段映射问题
- 优化欢迎语类型转换逻辑
- 添加通用工具函数（JSON、时间、字符串、切片处理）
- 更新ESLint配置和jsconfig配置

+++ api-server-go/internal/handler/dashboard/marketing/auto_pull.go
@@ -112,7 +112,7 @@ func (h *WorkRoomAutoPullHandler) Show(c *gin.Context) {
 	// 构建响应数据
 	responseData := ShowResponse{
 		QrcodeName:   item.QrcodeName,
-		IsVerified:   item.IsVerified,
+		IsVerified:   int(item.IsVerified),
 		LeadingWords: item.LeadingWords,
 		Employees:    employees,
 		Tags:         tags,
@@ -152,7 +152,7 @@ func (h *WorkRoomAutoPullHandler) Store(c *gin.Context) {
 	item := &model.WorkRoomAutoPull{
 		CorpID:       corpID,
 		QrcodeName:   req.QrcodeName,
-		IsVerified:   req.IsVerified,
+		IsVerified:   uint(req.IsVerified),
 		LeadingWords: req.LeadingWords,
 		Employees:    req.Employees,
 		Tags:         req.Tags,
